docs: (3/3) move docs in their own folder, fix file paths
