+++ src/docs/old-docs-list.js
@@ -1,6 +1,8 @@
 const path = require("path");
 const glob = require("glob");
 
-let fileVersions = glob.sync("./src/views/old-docs/*.mdx").map(v => path.parse(v).name);
+let fileVersions = glob
+  .sync("./src/docs/views/old-docs/*.mdx")
+  .map((v) => path.parse(v).name);
 
 module.exports = { fileVersions };
